更新Java项目组织名
