--- springboot/src/main/java/com/example/SpringbootApplication.java
@@ -1,15 +1,0 @@
-package com.example;
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-@MapperScan("com.example.mapper")
-public class SpringbootApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpringbootApplication.class, args);
-    }
-
-}
