--- springboot/src/main/java/com/example/common/Constants.java
@@ -1,9 +1,0 @@
-package com.example.common;
-
-public interface Constants {
-
-    String TOKEN = "token";
-
-    String USER_DEFAULT_PASSWORD = "123";
-
-}
