--- springboot/src/main/java/com/example/common/enums/RoleEnum.java
@@ -1,6 +1,0 @@
-package com.example.common.enums;
-
-public enum RoleEnum {
-    // 管理员
-    ADMIN,
-}
