--- springboot/src/main/java/com/example/controller/TypeController.java
@@ -1,87 +1,0 @@
-package com.example.controller;
-
-import com.example.common.Result;
-import com.example.entity.Type;
-import com.example.service.TypeService;
-import com.github.pagehelper.PageInfo;
-import org.springframework.web.bind.annotation.*;
-
-import javax.annotation.Resource;
-import java.util.List;
-
-/**
- * 公告信息表前端操作接口
- **/
-@RestController
-@RequestMapping("/type")
-public class TypeController {
-
-    @Resource
-    private TypeService typeService;
-
-    /**
-     * 新增
-     */
-    @PostMapping("/add")
-    public Result add(@RequestBody Type type) {
-        typeService.add(type);
-        return Result.success();
-    }
-
-    /**
-     * 删除
-     */
-    @DeleteMapping("/delete/{id}")
-    public Result deleteById(@PathVariable Integer id) {
-        typeService.deleteById(id);
-        return Result.success();
-    }
-
-    /**
-     * 批量删除
-     */
-    @DeleteMapping("/delete/batch")
-    public Result deleteBatch(@RequestBody List<Integer> ids) {
-        typeService.deleteBatch(ids);
-        return Result.success();
-    }
-
-    /**
-     * 修改
-     */
-    @PutMapping("/update")
-    public Result updateById(@RequestBody Type type) {
-        typeService.updateById(type);
-        return Result.success();
-    }
-
-    /**
-     * 根据ID查询
-     */
-    @GetMapping("/selectById/{id}")
-    public Result selectById(@PathVariable Integer id) {
-        Type type = typeService.selectById(id);
-        return Result.success(type);
-    }
-
-    /**
-     * 查询所有
-     */
-    @GetMapping("/selectAll")
-    public Result selectAll(Type type ) {
-        List<Type> list = typeService.selectAll(type);
-        return Result.success(list);
-    }
-
-    /**
-     * 分页查询
-     */
-    @GetMapping("/selectPage")
-    public Result selectPage(Type type,
-                             @RequestParam(defaultValue = "1") Integer pageNum,
-                             @RequestParam(defaultValue = "10") Integer pageSize) {
-        PageInfo<Type> page = typeService.selectPage(type, pageNum, pageSize);
-        return Result.success(page);
-    }
-
-}
