--- springboot/src/main/java/com/example/controller/WebController.java
@@ -1,73 +1,0 @@
-package com.example.controller;
-
-import cn.hutool.core.util.ObjectUtil;
-import cn.hutool.core.util.StrUtil;
-import com.example.common.Result;
-import com.example.common.enums.ResultCodeEnum;
-import com.example.common.enums.RoleEnum;
-import com.example.entity.Account;
-import com.example.service.AdminService;
-import org.springframework.web.bind.annotation.*;
-
-import javax.annotation.Resource;
-
-/**
- * 基础前端接口
- */
-@RestController
-public class WebController {
-
-    @Resource
-    private AdminService adminService;
-
-    @GetMapping("/")
-    public Result hello() {
-        return Result.success("访问成功");
-    }
-
-    /**
-     * 登录
-     */
-    @PostMapping("/login")
-    public Result login(@RequestBody Account account) {
-        if (ObjectUtil.isEmpty(account.getUsername()) || ObjectUtil.isEmpty(account.getPassword())
-                || ObjectUtil.isEmpty(account.getRole())) {
-            return Result.error(ResultCodeEnum.PARAM_LOST_ERROR);
-        }
-        if (RoleEnum.ADMIN.name().equals(account.getRole())) {
-            account = adminService.login(account);
-        }
-        return Result.success(account);
-    }
-
-    /**
-     * 注册
-     */
-    @PostMapping("/register")
-    public Result register(@RequestBody Account account) {
-        if (StrUtil.isBlank(account.getUsername()) || StrUtil.isBlank(account.getPassword())
-                || ObjectUtil.isEmpty(account.getRole())) {
-            return Result.error(ResultCodeEnum.PARAM_LOST_ERROR);
-        }
-        if (RoleEnum.ADMIN.name().equals(account.getRole())) {
-            adminService.register(account);
-        }
-        return Result.success();
-    }
-
-    /**
-     * 修改密码
-     */
-    @PutMapping("/updatePassword")
-    public Result updatePassword(@RequestBody Account account) {
-        if (StrUtil.isBlank(account.getUsername()) || StrUtil.isBlank(account.getPassword())
-                || ObjectUtil.isEmpty(account.getNewPassword())) {
-            return Result.error(ResultCodeEnum.PARAM_LOST_ERROR);
-        }
-        if (RoleEnum.ADMIN.name().equals(account.getRole())) {
-            adminService.updatePassword(account);
-        }
-        return Result.success();
-    }
-
-}
