--- springboot/src/main/java/com/example/entity/Notice.java
@@ -1,62 +1,0 @@
-package com.example.entity;
-
-import java.io.Serializable;
-
-/**
- * 公告信息表
-*/
-public class Notice implements Serializable {
-    private static final long serialVersionUID = 1L;
-
-    /** ID */
-    private Integer id;
-    /** 标题 */
-    private String title;
-    /** 内容 */
-    private String content;
-    /** 创建时间 */
-    private String time;
-    /** 创建人 */
-    private String user;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public String getTime() {
-        return time;
-    }
-
-    public void setTime(String time) {
-        this.time = time;
-    }
-
-    public String getUser() {
-        return user;
-    }
-
-    public void setUser(String user) {
-        this.user = user;
-    }
-
-}
