--- springboot/src/main/java/com/example/entity/Type.java
@@ -1,51 +1,0 @@
-package com.example.entity;
-
-import java.io.Serializable;
-
-/**
- * 商品分类信息表
-*/
-public class Type implements Serializable {
-    private static final long serialVersionUID = 1L;
-
-    /** ID */
-    private Integer id;
-    /** 分类名称 */
-    private String name;
-    /** 分类描述 */
-    private String description;
-    /** 分类图标 */
-    private String img;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getImg() {
-        return img;
-    }
-
-    public void setImg(String img) {
-        this.img = img;
-    }
-}
