+++ springboot/src/main/java/edu/whu/recirco/SpringbootApplication.java
@@ -1,0 +1,15 @@
+package edu.whu.recirco;
+
+import org.mybatis.spring.annotation.MapperScan;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@MapperScan("edu.whu.recirco.mapper")
+public class SpringbootApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringbootApplication.class, args);
+    }
+
+}
