+++ springboot/src/main/java/edu/whu/recirco/common/Constants.java
@@ -1,0 +1,9 @@
+package edu.whu.recirco.common;
+
+public interface Constants {
+
+    String TOKEN = "token";
+
+    String USER_DEFAULT_PASSWORD = "123";
+
+}
