+++ springboot/src/main/java/edu/whu/recirco/common/config/WebConfig.java
@@ -1,0 +1,24 @@
+package edu.whu.recirco.common.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import javax.annotation.Resource;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Resource
+    private JwtInterceptor jwtInterceptor;
+
+    // 加自定义拦截器JwtInterceptor，设置拦截规则
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(jwtInterceptor).addPathPatterns("/**")
+                .excludePathPatterns("/")
+                .excludePathPatterns("/login")
+                .excludePathPatterns("/register")
+                .excludePathPatterns("/files/**");
+    }
+}
