+++ springboot/src/main/java/edu/whu/recirco/common/enums/RoleEnum.java
@@ -1,0 +1,6 @@
+package edu.whu.recirco.common.enums;
+
+public enum RoleEnum {
+    // 管理员
+    ADMIN,
+}
