+++ springboot/src/main/java/edu/whu/recirco/controller/AdminController.java
@@ -1,0 +1,87 @@
+package edu.whu.recirco.controller;
+
+import edu.whu.recirco.common.Result;
+import edu.whu.recirco.entity.Admin;
+import edu.whu.recirco.service.AdminService;
+import com.github.pagehelper.PageInfo;
+import org.springframework.web.bind.annotation.*;
+
+import javax.annotation.Resource;
+import java.util.List;
+
+/**
+ * 管理员前端操作接口
+ **/
+@RestController
+@RequestMapping("/admin")
+public class AdminController {
+
+    @Resource
+    private AdminService adminService;
+
+    /**
+     * 新增
+     */
+    @PostMapping("/add")
+    public Result add(@RequestBody Admin admin) {
+        adminService.add(admin);
+        return Result.success();
+    }
+
+    /**
+     * 删除
+     */
+    @DeleteMapping("/delete/{id}")
+    public Result deleteById(@PathVariable Integer id) {
+        adminService.deleteById(id);
+        return Result.success();
+    }
+
+    /**
+     * 批量删除
+     */
+    @DeleteMapping("/delete/batch")
+    public Result deleteBatch(@RequestBody List<Integer> ids) {
+        adminService.deleteBatch(ids);
+        return Result.success();
+    }
+
+    /**
+     * 修改
+     */
+    @PutMapping("/update")
+    public Result updateById(@RequestBody Admin admin) {
+        adminService.updateById(admin);
+        return Result.success();
+    }
+
+    /**
+     * 根据ID查询
+     */
+    @GetMapping("/selectById/{id}")
+    public Result selectById(@PathVariable Integer id) {
+        Admin admin = adminService.selectById(id);
+        return Result.success(admin);
+    }
+
+    /**
+     * 查询所有
+     */
+    @GetMapping("/selectAll")
+    public Result selectAll(Admin admin) {
+        List<Admin> list = adminService.selectAll(admin);
+        return Result.success(list);
+    }
+
+    /**
+     * 分页查询
+     */
+    @GetMapping("/selectPage")
+    public Result selectPage(Admin admin,
+                             @RequestParam(defaultValue = "1") Integer pageNum,
+                             @RequestParam(defaultValue = "10") Integer pageSize) {
+        PageInfo<Admin> page = adminService.selectPage(admin, pageNum, pageSize);
+        return Result.success(page);
+    }
+
+}
