+++ springboot/src/main/java/edu/whu/recirco/controller/NoticeController.java
@@ -1,0 +1,87 @@
+package edu.whu.recirco.controller;
+
+import edu.whu.recirco.common.Result;
+import edu.whu.recirco.entity.Notice;
+import edu.whu.recirco.service.NoticeService;
+import com.github.pagehelper.PageInfo;
+import org.springframework.web.bind.annotation.*;
+
+import javax.annotation.Resource;
+import java.util.List;
+
+/**
+ * 公告信息表前端操作接口
+ **/
+@RestController
+@RequestMapping("/notice")
+public class NoticeController {
+
+    @Resource
+    private NoticeService noticeService;
+
+    /**
+     * 新增
+     */
+    @PostMapping("/add")
+    public Result add(@RequestBody Notice notice) {
+        noticeService.add(notice);
+        return Result.success();
+    }
+
+    /**
+     * 删除
+     */
+    @DeleteMapping("/delete/{id}")
+    public Result deleteById(@PathVariable Integer id) {
+        noticeService.deleteById(id);
+        return Result.success();
+    }
+
+    /**
+     * 批量删除
+     */
+    @DeleteMapping("/delete/batch")
+    public Result deleteBatch(@RequestBody List<Integer> ids) {
+        noticeService.deleteBatch(ids);
+        return Result.success();
+    }
+
+    /**
+     * 修改
+     */
+    @PutMapping("/update")
+    public Result updateById(@RequestBody Notice notice) {
+        noticeService.updateById(notice);
+        return Result.success();
+    }
+
+    /**
+     * 根据ID查询
+     */
+    @GetMapping("/selectById/{id}")
+    public Result selectById(@PathVariable Integer id) {
+        Notice notice = noticeService.selectById(id);
+        return Result.success(notice);
+    }
+
+    /**
+     * 查询所有
+     */
+    @GetMapping("/selectAll")
+    public Result selectAll(Notice notice) {
+        List<Notice> list = noticeService.selectAll(notice);
+        return Result.success(list);
+    }
+
+    /**
+     * 分页查询
+     */
+    @GetMapping("/selectPage")
+    public Result selectPage(Notice notice,
+                             @RequestParam(defaultValue = "1") Integer pageNum,
+                             @RequestParam(defaultValue = "10") Integer pageSize) {
+        PageInfo<Notice> page = noticeService.selectPage(notice, pageNum, pageSize);
+        return Result.success(page);
+    }
+
+}
