+++ springboot/src/main/java/edu/whu/recirco/controller/WebController.java
@@ -1,0 +1,73 @@
+package edu.whu.recirco.controller;
+
+import cn.hutool.core.util.ObjectUtil;
+import cn.hutool.core.util.StrUtil;
+import edu.whu.recirco.common.Result;
+import edu.whu.recirco.common.enums.ResultCodeEnum;
+import edu.whu.recirco.common.enums.RoleEnum;
+import edu.whu.recirco.entity.Account;
+import edu.whu.recirco.service.AdminService;
+import org.springframework.web.bind.annotation.*;
+
+import javax.annotation.Resource;
+
+/**
+ * 基础前端接口
+ */
+@RestController
+public class WebController {
+
+    @Resource
+    private AdminService adminService;
+
+    @GetMapping("/")
+    public Result hello() {
+        return Result.success("访问成功");
+    }
+
+    /**
+     * 登录
+     */
+    @PostMapping("/login")
+    public Result login(@RequestBody Account account) {
+        if (ObjectUtil.isEmpty(account.getUsername()) || ObjectUtil.isEmpty(account.getPassword())
+                || ObjectUtil.isEmpty(account.getRole())) {
+            return Result.error(ResultCodeEnum.PARAM_LOST_ERROR);
+        }
+        if (RoleEnum.ADMIN.name().equals(account.getRole())) {
+            account = adminService.login(account);
+        }
+        return Result.success(account);
+    }
+
+    /**
+     * 注册
+     */
+    @PostMapping("/register")
+    public Result register(@RequestBody Account account) {
+        if (StrUtil.isBlank(account.getUsername()) || StrUtil.isBlank(account.getPassword())
+                || ObjectUtil.isEmpty(account.getRole())) {
+            return Result.error(ResultCodeEnum.PARAM_LOST_ERROR);
+        }
+        if (RoleEnum.ADMIN.name().equals(account.getRole())) {
+            adminService.register(account);
+        }
+        return Result.success();
+    }
+
+    /**
+     * 修改密码
+     */
+    @PutMapping("/updatePassword")
+    public Result updatePassword(@RequestBody Account account) {
+        if (StrUtil.isBlank(account.getUsername()) || StrUtil.isBlank(account.getPassword())
+                || ObjectUtil.isEmpty(account.getNewPassword())) {
+            return Result.error(ResultCodeEnum.PARAM_LOST_ERROR);
+        }
+        if (RoleEnum.ADMIN.name().equals(account.getRole())) {
+            adminService.updatePassword(account);
+        }
+        return Result.success();
+    }
+
+}
