+++ springboot/src/main/java/edu/whu/recirco/mapper/AdminMapper.java
@@ -1,0 +1,40 @@
+package edu.whu.recirco.mapper;
+
+import edu.whu.recirco.entity.Admin;
+import org.apache.ibatis.annotations.Select;
+
+import java.util.List;
+
+/**
+ * 操作admin相关数据接口
+ */
+public interface AdminMapper {
+
+    /**
+     * 新增
+     */
+    int insert(Admin admin);
+
+    /**
+     * 删除
+     */
+    int deleteById(Integer id);
+
+    /**
+     * 修改
+     */
+    int updateById(Admin admin);
+
+    /**
+     * 根据ID查询
+     */
+    Admin selectById(Integer id);
+
+    /**
+     * 查询所有
+     */
+    List<Admin> selectAll(Admin admin);
+
+    @Select("select * from admin where username = #{username}")
+    Admin selectByUsername(String username);
+}
