+++ springboot/src/main/java/edu/whu/recirco/mapper/TypeMapper.java
@@ -1,0 +1,37 @@
+package edu.whu.recirco.mapper;
+
+import edu.whu.recirco.entity.Type;
+
+import java.util.List;
+
+/**
+ * 操作notice相关数据接口
+ */
+public interface TypeMapper {
+
+    /**
+     * 新增
+     */
+    int insert(Type type);
+
+    /**
+     * 删除
+     */
+    int deleteById(Integer id);
+
+    /**
+     * 修改
+     */
+    int updateById(Type type);
+
+    /**
+     * 根据ID查询
+     */
+    Type selectById(Integer id);
+
+    /**
+     * 查询所有
+     */
+    List<Type> selectAll(Type type);
+
+}
