+++ springboot/src/main/resources/mapper/NoticeMapper.xml
@@ -2,57 +2,59 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.mapper.NoticeMapper">
+<mapper namespace="edu.whu.recirco.mapper.NoticeMapper">
 
     <sql id="Base_Column_List">
-        id,title,content,time,user
+        id
+        ,title,content,time,user
     </sql>
 
-    <select id="selectAll" resultType="com.example.entity.Notice">
+    <select id="selectAll" resultType="edu.whu.recirco.entity.Notice">
         select
-        <include refid="Base_Column_List" />
+        <include refid="Base_Column_List"/>
         from notice
         <where>
-            <if test="id != null"> and id= #{id}</if>
-            <if test="title != null"> and title like concat('%', #{title}, '%')</if>
-            <if test="content != null"> and content= #{content}</if>
-            <if test="time != null"> and time= #{time}</if>
-            <if test="user != null"> and user= #{user}</if>
+            <if test="id != null">and id= #{id}</if>
+            <if test="title != null">and title like concat('%', #{title}, '%')</if>
+            <if test="content != null">and content= #{content}</if>
+            <if test="time != null">and time= #{time}</if>
+            <if test="user != null">and user= #{user}</if>
         </where>
         order by id desc
     </select>
 
-    <select id="selectById" resultType="com.example.entity.Notice">
+    <select id="selectById" resultType="edu.whu.recirco.entity.Notice">
         select
-        <include refid="Base_Column_List" />
+        <include refid="Base_Column_List"/>
         from notice
         where id = #{id}
     </select>
 
     <delete id="deleteById">
-        delete from notice
-        where  id = #{id}
+        delete
+        from notice
+        where id = #{id}
     </delete>
 
-    <insert id="insert" parameterType="com.example.entity.Notice" useGeneratedKeys="true">
+    <insert id="insert" parameterType="edu.whu.recirco.entity.Notice" useGeneratedKeys="true">
         insert into notice
         <trim prefix="(" suffix=")" suffixOverrides=",">
-                <if test="id != null">id,</if>
-                <if test="title != null">title,</if>
-                <if test="content != null">content,</if>
-                <if test="time != null">time,</if>
-                <if test="user != null">user,</if>
+            <if test="id != null">id,</if>
+            <if test="title != null">title,</if>
+            <if test="content != null">content,</if>
+            <if test="time != null">time,</if>
+            <if test="user != null">user,</if>
         </trim>
         <trim prefix="values (" suffix=")" suffixOverrides=",">
-                <if test="id != null">#{id},</if>
-                <if test="title != null">#{title},</if>
-                <if test="content != null">#{content},</if>
-                <if test="time != null">#{time},</if>
-                <if test="user != null">#{user},</if>
+            <if test="id != null">#{id},</if>
+            <if test="title != null">#{title},</if>
+            <if test="content != null">#{content},</if>
+            <if test="time != null">#{time},</if>
+            <if test="user != null">#{user},</if>
         </trim>
     </insert>
 
-    <update id="updateById" parameterType="com.example.entity.Notice">
+    <update id="updateById" parameterType="edu.whu.recirco.entity.Notice">
         update notice
         <set>
             <if test="title != null">
@@ -68,7 +70,7 @@
                 user = #{user},
             </if>
         </set>
-        where id = #{id} 
+        where id = #{id}
     </update>
 
 </mapper>
