+++ springboot/src/main/resources/mapper/TypeMapper.xml
@@ -2,54 +2,56 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.mapper.TypeMapper">
+<mapper namespace="edu.whu.recirco.mapper.TypeMapper">
 
     <sql id="Base_Column_List">
-        id,name,description,img
+        id
+        ,name,description,img
     </sql>
 
-    <select id="selectAll" resultType="com.example.entity.Type">
+    <select id="selectAll" resultType="edu.whu.recirco.entity.Type">
         select
-        <include refid="Base_Column_List" />
+        <include refid="Base_Column_List"/>
         from type
         <where>
-            <if test="id != null"> and id= #{id}</if>
-            <if test="name != null"> and name like concat('%', #{name}, '%')</if>
-            <if test="description != null"> and description= #{description}</if>
-            <if test="img != null"> and img= #{img}</if>
+            <if test="id != null">and id= #{id}</if>
+            <if test="name != null">and name like concat('%', #{name}, '%')</if>
+            <if test="description != null">and description= #{description}</if>
+            <if test="img != null">and img= #{img}</if>
         </where>
         order by id desc
     </select>
 
-    <select id="selectById" resultType="com.example.entity.Type">
+    <select id="selectById" resultType="edu.whu.recirco.entity.Type">
         select
-        <include refid="Base_Column_List" />
+        <include refid="Base_Column_List"/>
         from type
         where id = #{id}
     </select>
 
     <delete id="deleteById">
-        delete from type
-        where  id = #{id}
+        delete
+        from type
+        where id = #{id}
     </delete>
 
-    <insert id="insert" parameterType="com.example.entity.Type" useGeneratedKeys="true">
+    <insert id="insert" parameterType="edu.whu.recirco.entity.Type" useGeneratedKeys="true">
         insert into type
         <trim prefix="(" suffix=")" suffixOverrides=",">
-                <if test="id != null">id,</if>
-                <if test="name != null">name,</if>
-                <if test="description != null">description,</if>
-                <if test="img != null">img,</if>
+            <if test="id != null">id,</if>
+            <if test="name != null">name,</if>
+            <if test="description != null">description,</if>
+            <if test="img != null">img,</if>
         </trim>
         <trim prefix="values (" suffix=")" suffixOverrides=",">
-                <if test="id != null">#{id},</if>
-                <if test="name != null">#{name},</if>
-                <if test="description != null">#{description},</if>
-                <if test="img != null">#{img},</if>
+            <if test="id != null">#{id},</if>
+            <if test="name != null">#{name},</if>
+            <if test="description != null">#{description},</if>
+            <if test="img != null">#{img},</if>
         </trim>
     </insert>
 
-    <update id="updateById" parameterType="com.example.entity.Type">
+    <update id="updateById" parameterType="edu.whu.recirco.entity.Type">
         update type
         <set>
             <if test="name != null">
@@ -62,7 +64,7 @@
                 img = #{img},
             </if>
         </set>
-        where id = #{id} 
+        where id = #{id}
     </update>
 
 </mapper>
